Notification update - vote
Busy gives notifications for some upvotes. Mostly Downvotes as far as I can tell. Corrected it. They did not show up when  I made this. So updated databases, and enums to support new type. Update ui and notification codes to give out the correct notifications for the new types and not crash the app. Incremented Db version So you can start afresh.