View notifications
A lot of work was put in to rehaul the way reply,mentions,vote and reblog notifications were displayed as well as how they open their posts. After digging through a lot of source code of steem I found out how they do it and added the functionality here .Added a new api call to facilitate the call. Which makes the whole process much faster instead of the three calls I would make in the past.