All the classes for search
All the classes required to perform a search also including android xml's for getting search to work. The activity is a singleton to avoid many of them, you can search from the same activity, results are displayed in the same page. For searching for users we use steemit and for posts we use asksteem.