bug fix
removed unique tag from the drafts db, it would not save others.
the loading bar would not stop, added a funtion to stop it.
sometimes context would be null so put a check for that
Main tags did not implement the app theme
the comment page did not implement the app theme entirely because android would not let the cardview go before the input. the text would be white and so would the background be. It is implemented correctly now by using the background color on my own as the windows would not show it and using the input as the background.