Search activity
Search activity, search suggestions provider, updates view holder for users and a new view holder for headers, Rpc call for fetching users and processing the search users returned.
Also, a bug fix , for the condition where some posts do not contain the tags as an array but as a string. this would cause the app to crash