UI changes
Changes ui in article preview, making it cleaner and more robust.
If the tag is too long, it will not make the app look ugly. Added an icon instead of the html unicode character
added the search icon
both as svg's and use theme colours
added the search menu item