FabHider class
the fab hider class will hide the fab button given to it on scroll or either the recyclerview, scrollview or a nested view.
For the button of the type FABmenu it will hide the view as it is not a true fab and does not have the fav hide button.