Mainactivity changes
typing in your physical keyboard when on the main activity will open the search bar automatically.
If text is shared and contains words, then it is sent to the post activity and saved as a draft. If not then it is checked for a url.
the search is opened if you click on the search bar by calling a system function.