Add widget support.
The widget provider with all the code to recieve intent broadcasts and upate the ui of the widget and all widgets. The widget xml file for declaring a widget. A widget service so list views can update and show the ui under a widget. Treated as remote views. The configuration acitivity which fires up as you select the widget is where you can follow a tag or select your feed. It also serves as the common place for all widgets to save their data, including the LG smart bulletin one. It also serves as a setting to change the widget into another one.

The misc constants have the code to compress and cut out a bitmap so It can be used in the widget without throwing the oom exceptions for widgets. The image looks bad but the app just works.