Add LG smart bulletin support
This will declare the widget for the lg smart bulletin. The differences being that it uses a different value instead of home and the widget is not of a fixed size. Hence we add the views to the widget and do not use a list view as it will not scroll. It uses the same configuration activity as the other widgets and saves its data there as well. The same bitmap scaling functions are also utilized.