add database support for the savetime field and fix a critical bug
The functions in the article repo now return back the updated savetime which is saved in the viewmodel so it remains even when orientation changes. Adds the methods to sort with it and updates all lists to also return the new field. Adds the update function to the dao's and now inserts will not replace but update.