add camera and gallery options for the user to choose from while uploading an image
the app now asks if camera or gallery is to be used while upload an image. easy permissions are used to handle permission support. After selecting/clicking an image the Ucrop editor will popup for them to edit the image as they see fit then passed back to our app for usage.