fix bugs brought by new steemit json results
the api's now have different json responses. A lot of functions added for easy usage across the app. Also to prevent most future crashes because of this. Where time is not present will use empty string to not induce erros. net_votes is now calculated by the length of the active votes array.