make processing of posts in async tasks, return the displayImage property for easier access
the processing of posts would happen on the ui thread before. now it will happen within async tasks in a background thread. This make the application smoother.
To reduce load a lot of json arrays were dropped in the last update, this will return a new element from the db so we only have one image to show.